+++ lib/getch/options.rb
@@ -2,7 +2,7 @@ require 'optparse'
 
 module Getch
   class Options
-    attr_reader :language, :zoneinfo, :keymap, :disk, :fs, :username, :encrypt, :verbose
+    attr_reader :language, :zoneinfo, :keymap, :disk, :fs, :username, :boot_disk, :swap_disk, :encrypt, :verbose
 
     def initialize(argv)
       @language = DEFAULT_OPTIONS[:language]
@@ -11,6 +11,8 @@ module Getch
       @disk = DEFAULT_OPTIONS[:disk]
       @fs = DEFAULT_OPTIONS[:fs]
       @username = DEFAULT_OPTIONS[:username]
+      @boot_disk = DEFAULT_OPTIONS[:boot_disk]
+      @swap_disk = DEFAULT_OPTIONS[:swap_disk]
       @encrypt = DEFAULT_OPTIONS[:encrypt]
       @verbose = DEFAULT_OPTIONS[:verbose]
       parse(argv)
@@ -29,7 +31,7 @@ module Getch
         opts.on("-k", "--keymap KEY", "Default is us") do |key|
           @keymap = key
         end
-        opts.on("-d", "--disk DISK", "Disk where install Gentoo (sda,sdb)") do |disk|
+        opts.on("-d", "--disk DISK", "Disk where install Gentoo (sda,sdb), default use #{@disk}") do |disk|
           @disk = disk
         end
         opts.on("-f", "--format FS", "Can be ext4, lvm or zfs. Default use ext4") do |fs|
@@ -38,6 +40,14 @@ module Getch
         end
         opts.on("-u", "--username USERNAME", "Create a partition /home and add a new user /home/USERNAME") do |user|
           @username = user
+        end
+        opts.on("--separate-boot DISK", "Disk for the boot partition, default use #{@disk}") do |boot|
+          @boot_disk = boot
+          DEFAULT_OPTIONS[:boot_disk] = boot
+        end
+        opts.on("--separate-swap DISK", "Disk for the swap partition, default use #{@disk}") do |swap|
+          @swap_disk = swap
+          DEFAULT_OPTIONS[:swap_disk] = swap
         end
         opts.on("--encrypt", "Encrypt your system.") do
           @encrypt = true